Allow objects with overloaded stringification to clone themselves.
By checking early for a clone_self() method in an eval block objects with overloaded stringification it is made possible for objects with overloaded stringification to clone themselves.  This partially fixes rt.cpan 18261.

I didn't add any tests since it isn't clear to me how to test this without using overload or Scalar::Util, and because quite frankly the existing tests are scary.